+++ Sources/PointMass.swift
@@ -60,14 +60,14 @@ public final class PointMass: Codable, VectorRepresentable {
     /// Averages a list of point mass positions into one normalized Vector2 point
     @_specialize(where T == Array<PointMass>)
     @_specialize(where T == ContiguousArray<PointMass>)
-    public static func averagePosition<T: Collection>(of pointMasses: T) -> Vector2 where T.Iterator.Element == PointMass, T.IndexDistance == Int {
+    public static func averagePosition<T: Collection>(of pointMasses: T) -> Vector2 where T.Iterator.Element == PointMass {
         return pointMasses.averageVector()
     }
     
     /// Averages a list of point mass velocities into one normalized Vector2 point
     @_specialize(where T == Array<PointMass>)
     @_specialize(where T == ContiguousArray<PointMass>)
-    public static func averageVelocity<T: Collection>(of pointMasses: T) -> Vector2 where T.Iterator.Element == PointMass, T.IndexDistance == Int {
+    public static func averageVelocity<T: Collection>(of pointMasses: T) -> Vector2 where T.Iterator.Element == PointMass {
         return pointMasses.reduce(Vector2.zero) { $0 + $1.velocity } / JFloat(pointMasses.count)
     }
 }